requirements.txt
lmdb
pandas
numpy
numpy-quaternion
opencv-python
tensorboardX
tensorboard
matplotlib
numba
scipy
